// src/models/users/users-type.ts
import {ObjectId} from "mongodb";

// export type TUserDb = {
//     _id: ObjectId
//     id: string
//     login: string
//     email: string
//     createdAt: string
//     passwordHash: string
//     passwordSalt: string
// }


export type TUserView = {
    id: string
    login: string
    email: string
    createdAt: string
}